ref(trace): more tests
