+++ static/app/views/performance/newTraceDetails/trace.spec.tsx
@@ -18,7 +18,7 @@ import type {RawSpanType} from 'sentry/components/events/interfaces/spans/types'
 import {EntryType, type Event, type EventTransaction} from 'sentry/types';
 import type {TraceFullDetailed} from 'sentry/utils/performance/quickTrace/types';
 import {TraceView} from 'sentry/views/performance/newTraceDetails/index';
-import type {TraceTree} from 'sentry/views/performance/newTraceDetails/traceTree';
+import type {TraceTree} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
 import {RouteContext} from 'sentry/views/routeContext';
 
 jest.mock('screenfull', () => ({
