+++ static/app/views/performance/newTraceDetails/traceDrawer/details/error.tsx
@@ -19,7 +19,11 @@ import type {TraceTreeNodeDetailsProps} from 'sentry/views/performance/newTraceD
 import {getTraceTabTitle} from 'sentry/views/performance/newTraceDetails/traceState/traceTabs';
 import {Row, Tags} from 'sentry/views/performance/traceDetails/styles';
 
-import {makeTraceNodeBarColor, type TraceTree, type TraceTreeNode} from '../../traceTree';
+import {
+  makeTraceNodeBarColor,
+  type TraceTree,
+  type TraceTreeNode,
+} from '../../traceModels/traceTree';
 
 import {IssueList} from './issues/issues';
 import {TraceDrawerComponents} from './styles';
