+++ static/app/views/performance/newTraceDetails/traceDrawer/details/noData.tsx
@@ -10,7 +10,7 @@ import type {TraceTreeNodeDetailsProps} from 'sentry/views/performance/newTraceD
 import {
   makeTraceNodeBarColor,
   type NoDataNode,
-} from 'sentry/views/performance/newTraceDetails/traceTree';
+} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
 import {Row} from 'sentry/views/performance/traceDetails/styles';
 
 export function NoDataDetails(props: TraceTreeNodeDetailsProps<NoDataNode>) {
