+++ static/app/views/performance/newTraceDetails/traceDrawer/details/siblingAutogroup.tsx
@@ -8,7 +8,10 @@ import type {TraceTreeNodeDetailsProps} from 'sentry/views/performance/newTraceD
 import {getTraceTabTitle} from 'sentry/views/performance/newTraceDetails/traceState/traceTabs';
 import {Row} from 'sentry/views/performance/traceDetails/styles';
 
-import {makeTraceNodeBarColor, type SiblingAutogroupNode} from '../../traceTree';
+import {
+  makeTraceNodeBarColor,
+  type SiblingAutogroupNode,
+} from '../../traceModels/traceTree';
 
 import {IssueList} from './issues/issues';
 import {TraceDrawerComponents} from './styles';
