+++ static/app/views/performance/newTraceDetails/traceModels/traceTree.tsx
@@ -22,9 +22,8 @@ import {
 } from 'sentry/utils/performance/vitals/constants';
 import type {Vital} from 'sentry/utils/performance/vitals/types';
 
-import {TraceType} from '../traceDetails/newTraceDetailsContent';
-import {isRootTransaction} from '../traceDetails/utils';
-
+import {TraceType} from '../../traceDetails/newTraceDetailsContent';
+import {isRootTransaction} from '../../traceDetails/utils';
 import {
   isAutogroupedNode,
   isMissingInstrumentationNode,
@@ -37,7 +36,7 @@ import {
   isTraceNode,
   isTransactionNode,
   shouldAddMissingInstrumentationSpan,
-} from './guards';
+} from '../guards';
 
 /**
  *
