+++ static/app/views/performance/newTraceDetails/traceRenderers/virtualizedViewManager.tsx
@@ -1,4 +1,3 @@
-import {useLayoutEffect, useRef, useState} from 'react';
 import * as Sentry from '@sentry/react';
 import {mat3, vec2} from 'gl-matrix';
 
@@ -6,8 +5,6 @@ import type {Client} from 'sentry/api';
 import type {Organization} from 'sentry/types';
 import {getDuration} from 'sentry/utils/formatters';
 import clamp from 'sentry/utils/number/clamp';
-import {requestAnimationTimeout} from 'sentry/utils/profiling/hooks/useVirtualizedTree/virtualizedTreeUtils';
-import {lightTheme as theme} from 'sentry/utils/theme';
 import {
   isAutogroupedNode,
   isMissingInstrumentationNode,
@@ -22,7 +19,10 @@ import {
 import {
   type TraceTree,
   TraceTreeNode,
-} from 'sentry/views/performance/newTraceDetails/traceTree';
+} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
+import {TraceRowWidthMeasurer} from 'sentry/views/performance/newTraceDetails/traceRenderers/traceRowWidthMeasurer';
+import {TraceTextMeasurer} from 'sentry/views/performance/newTraceDetails/traceRenderers/traceTextMeasurer';
+import {TraceView} from 'sentry/views/performance/newTraceDetails/traceRenderers/traceView';
 
 const DIVIDER_WIDTH = 6;
 
@@ -37,72 +37,8 @@ type ViewColumn = {
   width: number;
 };
 
-class View {
-  public x: number;
-  public y: number;
-  public width: number;
-  public height: number;
-
-  constructor(x: number, y: number, width: number, height: number) {
-    this.x = x;
-    this.y = y;
-    this.width = width;
-    this.height = height;
-  }
-
-  static From(view: View): View {
-    return new View(view.x, view.y, view.width, view.height);
-  }
-  static Empty(): View {
-    return new View(0, 0, 1000, 1);
-  }
-
-  serialize() {
-    return [this.x, this.y, this.width, this.height];
-  }
-
-  between(to: View): mat3 {
-    return mat3.fromValues(
-      to.width / this.width,
-      0,
-      0,
-      to.height / this.height,
-      0,
-      0,
-      to.x - this.x * (to.width / this.width),
-      to.y - this.y * (to.height / this.height),
-      1
-    );
-  }
-
-  transform(mat: mat3): [number, number, number, number] {
-    const x = this.x * mat[0] + this.y * mat[3] + mat[6];
-    const y = this.x * mat[1] + this.y * mat[4] + mat[7];
-    const width = this.width * mat[0] + this.height * mat[3];
-    const height = this.width * mat[1] + this.height * mat[4];
-    return [x, y, width, height];
-  }
-
-  get center() {
-    return this.x + this.width / 2;
-  }
-
-  get left() {
-    return this.x;
-  }
-  get right() {
-    return this.x + this.width;
-  }
-  get top() {
-    return this.y;
-  }
-  get bottom() {
-    return this.y + this.height;
-  }
-}
-
 export function computeTimelineIntervals(
-  view: View,
+  view: TraceView,
   targetInterval: number,
   results: (number | undefined)[]
 ): void {
@@ -154,26 +90,26 @@ export class VirtualizedViewManager {
   // Represents the space of the entire trace, for example
   // a trace starting at 0 and ending at 1000 would have a space of [0, 1000]
   to_origin: number = 0;
-  trace_space: View = View.Empty();
+  trace_space: TraceView = TraceView.Empty();
   // The view defines what the user is currently looking at, it is a subset
   // of the trace space. For example, if the user is currently looking at the
   // trace from 500 to 1000, the view would be represented by [x, width] = [500, 500]
-  trace_view: View = View.Empty();
+  trace_view: TraceView = TraceView.Empty();
   // Represents the pixel space of the entire trace - this is the container
   // that we render to. For example, if the container is 1000px wide, the
   // pixel space would be [0, 1000]
-  trace_physical_space: View = View.Empty();
-  container_physical_space: View = View.Empty();
+  trace_physical_space: TraceView = TraceView.Empty();
+  container_physical_space: TraceView = TraceView.Empty();
 
   events: EventStore = {
     ['divider resize end']: new Set<VirtualizedViewManagerEvents['divider resize end']>(),
   };
 
-  row_measurer: DOMWidthMeasurer<TraceTreeNode<TraceTree.NodeValue>> =
-    new DOMWidthMeasurer();
-  indicator_label_measurer: DOMWidthMeasurer<TraceTree['indicators'][0]> =
-    new DOMWidthMeasurer();
-  text_measurer: TextMeasurer = new TextMeasurer();
+  row_measurer: TraceRowWidthMeasurer<TraceTreeNode<TraceTree.NodeValue>> =
+    new TraceRowWidthMeasurer();
+  indicator_label_measurer: TraceRowWidthMeasurer<TraceTree['indicators'][0]> =
+    new TraceRowWidthMeasurer();
+  text_measurer: TraceTextMeasurer = new TraceTextMeasurer();
 
   resize_observer: ResizeObserver | null = null;
   list: VirtualizedList | null = null;
@@ -285,16 +221,16 @@ export class VirtualizedViewManager {
   initializeTraceSpace(space: [x: number, y: number, width: number, height: number]) {
     this.to_origin = space[0];
 
-    this.trace_space = new View(0, 0, space[2], space[3]);
-    this.trace_view = new View(0, 0, space[2], space[3]);
+    this.trace_space = new TraceView(0, 0, space[2], space[3]);
+    this.trace_view = new TraceView(0, 0, space[2], space[3]);
 
     this.recomputeTimelineIntervals();
     this.recomputeSpanToPxMatrix();
   }
 
   initializePhysicalSpace(width: number, height: number) {
-    this.container_physical_space = new View(0, 0, width, height);
-    this.trace_physical_space = new View(
+    this.container_physical_space = new TraceView(0, 0, width, height);
+    this.trace_physical_space = new TraceView(
       0,
       0,
       width * this.columns.span_list.width,
@@ -1632,171 +1568,6 @@ export class VirtualizedViewManager {
   }
 }
 
-// The backing cache should be a proper LRU cache,
-// so we dont end up storing an infinite amount of elements
-class DOMWidthMeasurer<T> {
-  cache: Map<T, number> = new Map();
-
-  queue: [T, HTMLElement][] = [];
-  drainRaf: number | null = null;
-  max: number = 0;
-
-  constructor() {
-    this.drain = this.drain.bind(this);
-  }
-
-  listeners: Record<'max', Set<(max: number) => void>> = {
-    max: new Set(),
-  };
-
-  on(event: 'max', cb: (max: number) => void) {
-    this.listeners?.[event]?.add?.(cb);
-  }
-
-  off(event: 'max', cb: (max: number) => void) {
-    this.listeners?.[event]?.delete?.(cb);
-  }
-
-  dispatch(max: number) {
-    for (const listener of this.listeners.max) {
-      listener(max);
-    }
-  }
-
-  enqueueMeasure(node: T, element: HTMLElement) {
-    if (this.cache.has(node)) {
-      return;
-    }
-
-    this.queue.push([node, element]);
-
-    if (this.drainRaf !== null) {
-      window.cancelAnimationFrame(this.drainRaf);
-    }
-    this.drainRaf = window.requestAnimationFrame(this.drain);
-  }
-
-  drain() {
-    const startMax = this.max;
-    while (this.queue.length > 0) {
-      const next = this.queue.pop()!;
-      const width = this.measure(next[0], next[1]);
-      if (width > this.max) {
-        this.max = width;
-      }
-    }
-
-    if (this.max !== startMax) {
-      this.dispatch(this.max);
-    }
-  }
-
-  measure(node: T, element: HTMLElement): number {
-    const cache = this.cache.get(node);
-    if (cache !== undefined) {
-      return cache;
-    }
-
-    const rect = element.getBoundingClientRect();
-    this.cache.set(node, rect.width);
-    return rect.width;
-  }
-}
-
-// The backing cache should be a proper LRU cache,
-// so we dont end up storing an infinite amount of elements
-class TextMeasurer {
-  queue: string[] = [];
-  drainRaf: number | null = null;
-  cache: Map<string, number> = new Map();
-
-  number: number = 0;
-  dot: number = 0;
-  duration: Record<string, number> = {};
-
-  constructor() {
-    this.drain = this.drain.bind(this);
-
-    const canvas = document.createElement('canvas');
-    const ctx = canvas.getContext('2d');
-
-    if (!ctx) {
-      for (const duration of ['ns', 'ms', 's', 'm', 'min', 'h', 'd']) {
-        // If for some reason we fail to create a canvas context, we can
-        // use a fallback value for the durations. It shouldnt happen,
-        // but it's better to have a fallback than to crash the entire app.
-        // I've made a couple manual measurements to determine a good fallback
-        // and 6.5px per letter seems like a reasonable approximation.
-        const PX_PER_LETTER = 6.5;
-        this.duration[duration] = duration.length * PX_PER_LETTER;
-      }
-      return;
-    }
-
-    canvas.width = 50 * window.devicePixelRatio ?? 1;
-    canvas.height = 50 * window.devicePixelRatio ?? 1;
-
-    ctx.font = '11px' + theme.text.family;
-
-    this.dot = ctx.measureText('.').width;
-    for (let i = 0; i < 10; i++) {
-      const measurement = ctx.measureText(i.toString());
-      this.number = Math.max(this.number, measurement.width);
-    }
-
-    for (const duration of ['ns', 'ms', 's', 'm', 'min', 'h', 'd']) {
-      this.duration[duration] = ctx.measureText(duration).width;
-    }
-  }
-
-  drain() {
-    for (const string of this.queue) {
-      this.measure(string);
-    }
-  }
-
-  computeStringLength(string: string): number {
-    let width = 0;
-    for (let i = 0; i < string.length; i++) {
-      switch (string[i]) {
-        case '.':
-          width += this.dot;
-          break;
-        case '0':
-        case '1':
-        case '2':
-        case '3':
-        case '4':
-        case '5':
-        case '6':
-        case '7':
-        case '8':
-        case '9':
-          width += this.number;
-          break;
-        default:
-          const remaining = string.slice(i);
-          if (this.duration[remaining]) {
-            width += this.duration[remaining];
-            return width;
-          }
-      }
-    }
-    return width;
-  }
-
-  measure(string: string): number {
-    const cached_width = this.cache.get(string);
-    if (cached_width !== undefined) {
-      return cached_width;
-    }
-
-    const width = this.computeStringLength(string);
-    this.cache.set(string, width);
-    return width;
-  }
-}
-
 // Jest does not implement scroll updates, however since we have the
 // middleware to handle scroll updates, we can dispatch a scroll event ourselves
 function dispatchJestScrollUpdate(container: HTMLElement) {
@@ -1846,374 +1617,11 @@ export class VirtualizedList {
   }
 }
 
-function maybeToggleScrollbar(
-  container: HTMLElement,
-  containerHeight: number,
-  scrollHeight: number,
-  manager: VirtualizedViewManager
-) {
-  if (scrollHeight > containerHeight) {
-    container.style.overflowY = 'scroll';
-    container.style.scrollbarGutter = 'stable';
-    manager.onScrollbarWidthChange(container.offsetWidth - container.clientWidth);
-  } else {
-    container.style.overflowY = 'auto';
-    container.style.scrollbarGutter = 'auto';
-    manager.onScrollbarWidthChange(0);
-  }
-}
-
-interface UseVirtualizedListProps {
-  container: HTMLElement | null;
-  items: ReadonlyArray<TraceTreeNode<TraceTree.NodeValue>>;
-  manager: VirtualizedViewManager;
-  render: (item: VirtualizedRow) => React.ReactNode;
-}
-
-interface UseVirtualizedListResult {
-  list: VirtualizedList;
-  rendered: React.ReactNode[];
-  virtualized: VirtualizedRow[];
-}
-
-export const useVirtualizedList = (
-  props: UseVirtualizedListProps
-): UseVirtualizedListResult => {
-  const list = useRef<VirtualizedList | null>();
-
-  const scrollTopRef = useRef<number>(0);
-  const scrollHeightRef = useRef<number>(0);
-  const scrollContainerRef = useRef<HTMLElement | null>(null);
-
-  const renderCache = useRef<Map<number, React.ReactNode>>();
-  const styleCache = useRef<Map<number, React.CSSProperties>>();
-  const resizeObserverRef = useRef<ResizeObserver | null>(null);
-
-  if (!styleCache.current) {
-    styleCache.current = new Map();
-  }
-  if (!renderCache.current) {
-    renderCache.current = new Map();
-  }
-
-  const [items, setItems] = useState<{
-    rendered: React.ReactNode[];
-    virtualized: VirtualizedRow[];
-  }>({rendered: [], virtualized: []});
-
-  if (!list.current) {
-    list.current = new VirtualizedList();
-    props.manager.registerList(list.current);
-  }
-
-  const renderRef = useRef<(item: VirtualizedRow) => React.ReactNode>(props.render);
-  renderRef.current = props.render;
-  const itemsRef = useRef<ReadonlyArray<TraceTreeNode<TraceTree.NodeValue>>>(props.items);
-  itemsRef.current = props.items;
-  const managerRef = useRef<VirtualizedViewManager>(props.manager);
-  managerRef.current = props.manager;
-
-  useLayoutEffect(() => {
-    if (!props.container) {
-      return;
-    }
-    const scrollContainer = props.container.children[0] as HTMLElement | null;
-    if (!scrollContainer) {
-      throw new Error(
-        'Virtualized list container has to render a scroll container as its first child.'
-      );
-    }
-  }, [props.container, props.items.length]);
-
-  useLayoutEffect(() => {
-    if (!props.container || !list.current) {
-      return;
-    }
-
-    list.current.container = props.container;
-
-    if (resizeObserverRef.current) {
-      resizeObserverRef.current.disconnect();
-    }
-
-    const resizeObserver = new ResizeObserver(elements => {
-      // We only care about changes to the height of the scroll container,
-      // if it has not changed then do not update the scroll height.
-      styleCache.current?.clear();
-      renderCache.current?.clear();
-
-      scrollHeightRef.current = elements[0].contentRect.height;
-      if (list.current) {
-        list.current.scrollHeight = scrollHeightRef.current;
-      }
-
-      maybeToggleScrollbar(
-        elements[0].target as HTMLElement,
-        scrollHeightRef.current,
-        itemsRef.current.length * 24,
-        managerRef.current
-      );
-
-      const recomputedItems = findRenderedItems({
-        scrollTop: scrollTopRef.current,
-        items: itemsRef.current,
-        overscroll: 5,
-        rowHeight: 24,
-        scrollHeight: scrollHeightRef.current,
-        styleCache: styleCache.current!,
-        renderCache: renderCache.current!,
-        render: renderRef.current,
-        manager: managerRef.current,
-      });
-      setItems(recomputedItems);
-    });
-
-    resizeObserver.observe(props.container);
-    resizeObserverRef.current = resizeObserver;
-  }, [props.container]);
-
-  const rafId = useRef<number | null>(null);
-  const pointerEventsRaf = useRef<{id: number} | null>(null);
-
-  useLayoutEffect(() => {
-    if (!list.current || !props.container) {
-      return undefined;
-    }
-
-    if (props.container && !scrollContainerRef.current) {
-      scrollContainerRef.current = props.container.children[0] as HTMLElement | null;
-    }
-
-    props.container.style.height = '100%';
-    props.container.style.overflow = 'auto';
-    props.container.style.position = 'relative';
-    props.container.style.willChange = 'transform';
-    props.container.style.overscrollBehavior = 'none';
-
-    scrollContainerRef.current!.style.overflow = 'hidden';
-    scrollContainerRef.current!.style.position = 'relative';
-    scrollContainerRef.current!.style.willChange = 'transform';
-    scrollContainerRef.current!.style.height = `${props.items.length * 24}px`;
-
-    maybeToggleScrollbar(
-      props.container,
-      scrollHeightRef.current,
-      props.items.length * 24,
-      props.manager
-    );
-
-    const onScroll = event => {
-      if (!list.current) {
-        return;
-      }
-
-      if (rafId.current !== null) {
-        window.cancelAnimationFrame(rafId.current);
-      }
-
-      managerRef.current.scrolling_source = 'list';
-      managerRef.current.enqueueOnScrollEndOutOfBoundsCheck();
-
-      rafId.current = window.requestAnimationFrame(() => {
-        scrollTopRef.current = Math.max(0, event.target?.scrollTop ?? 0);
-
-        const recomputedItems = findRenderedItems({
-          scrollTop: scrollTopRef.current,
-          items: props.items,
-          overscroll: 5,
-          rowHeight: 24,
-          scrollHeight: scrollHeightRef.current,
-          styleCache: styleCache.current!,
-          renderCache: renderCache.current!,
-          render: renderRef.current,
-          manager: managerRef.current,
-        });
-        setItems(recomputedItems);
-      });
-
-      if (!pointerEventsRaf.current && scrollContainerRef.current) {
-        scrollContainerRef.current.style.pointerEvents = 'none';
-      }
-
-      if (pointerEventsRaf.current) {
-        window.cancelAnimationFrame(pointerEventsRaf.current.id);
-      }
-
-      pointerEventsRaf.current = requestAnimationTimeout(() => {
-        styleCache.current?.clear();
-        renderCache.current?.clear();
-
-        managerRef.current.scrolling_source = null;
-
-        const recomputedItems = findRenderedItems({
-          scrollTop: scrollTopRef.current,
-          items: props.items,
-          overscroll: 5,
-          rowHeight: 24,
-          scrollHeight: scrollHeightRef.current,
-          styleCache: styleCache.current!,
-          renderCache: renderCache.current!,
-          render: renderRef.current,
-          manager: managerRef.current,
-        });
-        setItems(recomputedItems);
-
-        if (list.current && scrollContainerRef.current) {
-          scrollContainerRef.current.style.pointerEvents = 'auto';
-          pointerEventsRaf.current = null;
-        }
-      }, 50);
-    };
-
-    props.container.addEventListener('scroll', onScroll, {passive: true});
-
-    return () => {
-      props.container?.removeEventListener('scroll', onScroll);
-    };
-  }, [props.container, props.items, props.items.length, props.manager]);
-
-  useLayoutEffect(() => {
-    if (!list.current || !styleCache.current || !renderCache.current) {
-      return;
-    }
-
-    styleCache.current.clear();
-    renderCache.current.clear();
-
-    const recomputedItems = findRenderedItems({
-      scrollTop: scrollTopRef.current,
-      items: props.items,
-      overscroll: 5,
-      rowHeight: 24,
-      scrollHeight: scrollHeightRef.current,
-      styleCache: styleCache.current!,
-      renderCache: renderCache.current,
-      render: renderRef.current,
-      manager: managerRef.current,
-    });
-
-    setItems(recomputedItems);
-  }, [props.items, props.items.length, props.render]);
-
-  return {
-    virtualized: items.virtualized,
-    rendered: items.rendered,
-    list: list.current!,
-  };
-};
-
 export interface VirtualizedRow {
   index: number;
   item: TraceTreeNode<TraceTree.NodeValue>;
   key: number;
   style: React.CSSProperties;
-}
-
-function findRenderedItems({
-  items,
-  overscroll,
-  rowHeight,
-  scrollHeight,
-  scrollTop,
-  styleCache,
-  renderCache,
-  render,
-  manager,
-}: {
-  items: ReadonlyArray<TraceTreeNode<TraceTree.NodeValue>>;
-  manager: VirtualizedViewManager;
-  overscroll: number;
-  render: (arg: VirtualizedRow) => React.ReactNode;
-  renderCache: Map<number, React.ReactNode>;
-  rowHeight: number;
-  scrollHeight: number;
-  scrollTop: number;
-  styleCache: Map<number, React.CSSProperties>;
-}): {rendered: React.ReactNode[]; virtualized: VirtualizedRow[]} {
-  // This is overscroll height for single direction, when computing the total,
-  // we need to multiply this by 2 because we overscroll in both directions.
-  const OVERSCROLL_HEIGHT = overscroll * rowHeight;
-  const virtualized: VirtualizedRow[] = [];
-  const rendered: React.ReactNode[] = [];
-
-  // Clamp viewport to scrollHeight bounds [0, length * rowHeight] because some browsers may fire
-  // scrollTop with negative values when the user scrolls up past the top of the list (overscroll behavior)
-  const viewport = {
-    top: Math.max(scrollTop - OVERSCROLL_HEIGHT, 0),
-    bottom: Math.min(
-      scrollTop + scrollHeight + OVERSCROLL_HEIGHT,
-      items.length * rowHeight
-    ),
-  };
-
-  // Points to the position inside the visible array
-  let visibleItemIndex = 0;
-  // Points to the currently iterated item
-  let indexPointer = findOptimisticStartIndex({
-    items,
-    viewport,
-    scrollTop,
-    rowHeight,
-    overscroll,
-  });
-
-  manager.start_virtualized_index = indexPointer;
-
-  // Max number of visible items in our list
-  const MAX_VISIBLE_ITEMS = Math.ceil((scrollHeight + OVERSCROLL_HEIGHT * 2) / rowHeight);
-  const ALL_ITEMS = items.length;
-
-  // While number of visible items is less than max visible items, and we haven't reached the end of the list
-  while (visibleItemIndex < MAX_VISIBLE_ITEMS && indexPointer < ALL_ITEMS) {
-    const elementTop = indexPointer * rowHeight;
-    const elementBottom = elementTop + rowHeight;
-
-    // An element is inside a viewport if the top of the element is below the top of the viewport
-    // and the bottom of the element is above the bottom of the viewport
-    if (elementTop >= viewport.top && elementBottom <= viewport.bottom) {
-      let style = styleCache.get(indexPointer);
-      if (!style) {
-        style = {position: 'absolute', top: elementTop};
-        styleCache.set(indexPointer, style);
-      }
-
-      const virtualizedRow: VirtualizedRow = {
-        key: indexPointer,
-        style,
-        index: indexPointer,
-        item: items[indexPointer],
-      };
-
-      virtualized[visibleItemIndex] = virtualizedRow;
-
-      const renderedRow = renderCache.get(indexPointer) || render(virtualizedRow);
-      rendered[visibleItemIndex] = renderedRow;
-      renderCache.set(indexPointer, renderedRow);
-      visibleItemIndex++;
-    }
-    indexPointer++;
-  }
-
-  return {rendered, virtualized};
-}
-
-export function findOptimisticStartIndex({
-  items,
-  overscroll,
-  rowHeight,
-  scrollTop,
-  viewport,
-}: {
-  items: ReadonlyArray<TraceTreeNode<TraceTree.NodeValue>>;
-  overscroll: number;
-  rowHeight: number;
-  scrollTop: number;
-  viewport: {bottom: number; top: number};
-}): number {
-  if (!items.length || viewport.top === 0) {
-    return 0;
-  }
-  return Math.max(Math.floor(scrollTop / rowHeight) - overscroll, 0);
 }
 
 function findInTreeFromSegment(
