+++ static/app/views/performance/newTraceDetails/traceState/traceTabs.tsx
@@ -1,11 +1,11 @@
 import * as Sentry from '@sentry/react';
 
 import {t} from 'sentry/locale';
-import {traceReducerExhaustiveActionCheck} from 'sentry/views/performance/newTraceDetails/traceState';
 import type {
   TraceTree,
   TraceTreeNode,
-} from 'sentry/views/performance/newTraceDetails/traceTree';
+} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
+import {traceReducerExhaustiveActionCheck} from 'sentry/views/performance/newTraceDetails/traceState';
 
 import {
   isAutogroupedNode,
